Add lesson on github for introduction
Besides wanting to bring students away from the software engineering dark-ages pre-git, we want to get students used to working with git, an essential tool in modern days for any software developer.
Moreover, in order to perform the classroom exercises and have reviews from TAs, we need students to be confortable with it.

Signed-off-by: diogoaurelio <diogopaurelio@gmail.com>